pyproject.toml
[tool.poetry]
name = "goit-homework-m2t12"
version = "0.1.0"
description = ""
authors = ["Roman-K26 <romankuspys777@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.23"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.37.0"
sphinx = "^7.2.6"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
